requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pymupdf==1.27.1
pillow==10.4.0
jinja2==3.1.4
python-multipart==0.0.9
pyyaml==6.0.2
anthropic==0.83.0
ollama==0.4.4
